llm-query-system/backend/requirements.txt
fastapi
uvicorn
requests
fitz  # PyMuPDF
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
